third_party/WebKit/LayoutTests/typedcssom/the-stylepropertymap/inline/getProperties.html

<!doctype html>
<meta charset="utf-8">
<title>StylePropertyMap.getProperties tests</title>
<link rel="help" href="https://drafts.css-houdini.org/css-typed-om-1/#dom-stylepropertymap-getproperties">
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<div id="target-empty"></div>
<div id="target-multiple" style="--A: A; width: 0px; --C: C; transition-duration: 1s, 2s; color: red; --B: B;"></div>
<script>
'use strict';

test(() => {
  const styleMap = document.getElementById('target-empty').attributeStyleMap;
  const properties = styleMap.getProperties();
  assert_equals(properties.length, 0);
}, 'Calling StylePropertyMap.getProperties on an empty property model returns a zero-length array');

test(() => {
  const styleMap = document.getElementById('target-multiple').attributeStyleMap;
  const properties = styleMap.getProperties();
  assert_array_equals(properties,
    ['color', 'transition-duration', 'width', '--A', '--B', '--C']);
}, 'StylePropertyMap.getProperties returns property names in correct order');

</script>
